client-side fuzzy search (#33)
openlibrary's search rankings have improved as of late, but there's still a risk of them returning something totally non-matching as the top result(s). so we can use our own fuzzy search to eliminate or demote bad results.

after the edition-aware results come back from openlibrary, we combine with the earlier openlibrary results, then apply a fuzzy-search ranking to re-rank all the results.

here, "apply" means each result is given a score of
```
original OL rank * fuzzy-search rank
```
and the results are sorted, lowest first.

results that fail the fuzzy search entirely are eliminated.

(the non-edition-aware results are not fuzzy searched because they might not be in english so there's no point.)

the records from our db continue to always come first, and are not affected by the fuzzy search. however, if we find that it would be better, we can change it to combine all the results first and apply fuzzy search on everything, OR make it "books from db, ranked with fuzzy search, followed by books from OL, ranked with fuzzy search".

also:
+ uses openlibrary's testing env, which hosts their latest search ranking logic
+ we now dedupe and apply the limit at the end of all of the above, not within each OL search call as we used to
+ actually, all of our own manipulations have been moved into a separate module, so that `OpenLibrary.search()` mostly just makes the api call and converts the results into `Book` objects with a minimum of filtering
+ updates the deduping logic to use a "looser" equality
+ updates the limit from 8 to 5, because we seem to almost always get a good result in top 5 now, yay
+ updates the debounce threshold from 300ms to 500ms. the point of 300ms is to make it feel super snappy, assuming your backend can keep up. but since OL can't respond in time to keep up with this rate, it just seems to cause more instability, like a chain of too many responses coming in late and messing up the results. with the 500ms, you have to be a bit more patient but you get better outcomes, I think/hope. we can keep tweaking this number. if it feels like the search box is ignoring you, we should lower it again.
+ we no longer preserve the order of the first set of OL results, as everything needs to be re-ranked with the fuzzy search anyway :( so you do get the disorienting change in order IF the fuzzy search affected things.