book notes (log a book) (#36)
creates book_notes table. this table is meant to be shared by 2 different concepts in the app:
1. a journal entry aka note (depending on context of where it is viewed): intended to be more personal, includes text, start date, finish date, and whether you finished the book or not (all optional). depending on what you filled out, these will later auto-assign the book to special lists like "read" or "did not finish".
2. a discussion post (eventually this starts a "thread"): intended to be more public, includes title, text, link_url (not all required).

as far as the user can see, these will be different objects, but they share so many behaviors that I think it makes sense for them to share a table and a type. hence the generic-sounding "book notes." they might be used for a reading timeline (using start date and finish date), or for user book reviews, or for sharing links to related articles and starting discussion threads.

this PR implements use case 1 above:
+ creates table
+ from a book's page, you can log a book, which creates a new book note
+ recent book notes are listed on a book's page, and also on a user's profile

not yet implemented in this PR because it got too big!:
+ edit and delete a note
+ "see all" pages (the "see all" links added in this PR don't work!)
+ UI nice-to-haves like truncating long texts with a "see more / see less" link, and a fancy checkbox

also:
+ extracts component `NameWithAvatar` to reuse wherever pieces of content have a "byline"

---

"log a book" modal:

![Screenshot 2023-12-05 at 9 46 32 PM](https://github.com/buidlers-fm/catalog/assets/4686089/b7453ff6-8489-475c-a375-2530752a26fb)

---

book page:

![Screenshot 2023-12-05 at 9 47 00 PM](https://github.com/buidlers-fm/catalog/assets/4686089/abd9ad10-561a-4dec-bc0e-61a63f874b8c)

---

user profile page:
![Screenshot 2023-12-05 at 9 47 32 PM](https://github.com/buidlers-fm/catalog/assets/4686089/d28c3b2f-e4cc-4f6f-9702-7bba8a8256cb)