redo "log a book" to support all reading statuses (#37)
I realized the feature as implemented in my last PR #36 assumed that you would log a book only upon finishing or abandoning it, but actually it should support logging upon starting, or just while still reading, as well. additionally, a "read" of a book should be a distinct thing, allowing multiple notes to be grouped together by belonging to the same "read."

to that end, this PR creates the `book_reads` table, which holds the start and end dates. (these date columns are now deprecated from the `book_notes` table and will be dropped later, along with the `finished` column). a BookNote now belongs to a BookRead. a BookNote now also has `readingStatus`, which can be `started`, `reading`, `finished`, or `abandoned`.

I've changed the "log this book" modal so that now the first step is to choose the reading status; then the form fields (and copy) are shown, because they differ slightly between the different statuses. if you choose any status other than `started`, it looks for the latest book read that has a start date but no end date, and populates the start date from that. otherwise it defaults dates to the current date.

you can use it in any number of ways:
+ log each step in the reading flow: started -> still reading (as many as you want) -> finished OR abandoned. this should all register as 1 book read.
+ log any step in the reading flow without first logging "started". this should create a new book read.

there is also now a status indicator on the book note card.

the reads and start/end dates aren't shown anywhere; they will go into a reading timeline-ish feature later on.

![log-a-book-pt-2](https://github.com/buidlers-fm/catalog/assets/4686089/07da836c-62cd-448e-ba4c-420bcaa14ee9)

also:
+ changed max text length from 10_000 to 1000 because we decided short is good

caveats:
+ I wasn't sure whether to require dates or not (when the field is present in the form). for now I've left it optional, but I'm not sure what the use cases will be, that would make people want to clear the date(s)-- because they want to log a read, just without specifying dates? because they DON'T want to log a read, and only want to write a note?-- so the behavior is probably a little wonky if you clear one or both of the dates. it assumes the former (log a new read), so if you create N book notes without dates, it will think you read the book N times. the answer is probably to support both use cases eventually and let the user specify which one they intend.