book link posts: CRUD (#40)
book posts are a separate section from book _notes_, and a post can be either a link post OR a text post, not both. (same as reddit, HN.) this PR is just for link posts, but text posts will be added fairly soon, so some of the code here is set up in preparation for that (e.g. looks for either type when fetching posts).

<img width="804" alt="Screenshot 2023-12-10 at 7 28 02 PM" src="https://github.com/buidlers-fm/catalog/assets/4686089/298eedd4-811f-498b-8953-d3b5f4d4e5fb">
---
<img width="794" alt="Screenshot 2023-12-10 at 7 28 12 PM" src="https://github.com/buidlers-fm/catalog/assets/4686089/0aa4beef-d2b3-4808-b57b-123a9ef53e66">


as all types still share a table, the logic for notes has been adjusted a bit to support these new types existing.

still struggling with the naming a bit: the table on the backend is called "book notes", but has 3 types:
+ journal_entry -> in UI is called "notes"
+ link_post -> in UI is called "posts", specifically "link posts"
+ text_post -> in UI (will be) called "posts", specifically "text posts"

so it's confusing that the word "notes" is overloaded here.. one or the other needs a new name, but I haven't thought of a good one yet.