upload cover images to supabase, pt 3 (#76)
in #69 I forgot to look at ALL places where books get created. when creating/updating lists, books are created in a batch; however, supabase doesn't allow batched file uploads. so it would be too slow to try to process the cover image for each book one by one while the user is waiting for a response. but it's not necessary because the processing is just an optimization. we can use the OL cover image url right away, then run a cron job that pings an endpoint to process N books at a time. (I've set it to 30 books at a time, once an hour, which is more than enough for our current needs, but can be easily tweaked as well.)

this approach should work until there are so many books to process that it exceeds something like
```
number of books that can be processed within the 60s vercel timeout x number of minutes in a day (1440)
```
(which would be like 50-100k books/day?)

other small things:
+ use `findOrCreateBook` when adding 1 book to a list, too - neglected in #69 
+ save thumbnail url if already available (ie from openlibrary)
+ in search results, fall back to large cover img if no thumbnail available
+ fix "add book to list" toast copy