fix: add external book to list (#141)
fixes reported bug where you can't add an external book to an existing list via "add to lists" modal. most likely introduced in #138. it was also obscured by the fact that the book page uses `RemountOnPathChange` which makes it non-obvious when the props have drifted out of sync, so that is also addressed here by explicitly specifying the expected props.