fix: notifs for follow and current status hidden (#170)
some previous PR fixed things so that a notif for a deleted object is never shown, ie, the notif requires its object to be present. however, it needed to take into account that when the object is a user (for "follow" notifs) or is a user current status (for "liked your current status" notifs) the object actually isn't needed and shouldn't be required.