fix: do process author if found by OL author id (#212)
the recently implemented duplicate person check was too early. if a new book's author is found in the db by their openLibraryAuthorId, they SHOULD go ahead and get linked, which doesn't create a new person. we should only run the duplicate person check if that is not the case, to make sure we don't create a duplicate person.